all CRUD working with exceptions being handled

+++ rest-common/src/main/java/service/DatasetService.java
@@ -64,6 +64,9 @@ public class DatasetService {
             entityManager.getTransaction().begin();
         }
         Datasets dataset = entityManager.find(Datasets.class, id);
+        if(dataset == null){
+            throw new NullPointerException("Dataset with id " + id + " not found");
+        }
         entityManager.remove(dataset);
         entityManager.getTransaction().commit();
     }
